+++ lib/spreedly/environment.rb
@@ -92,8 +92,8 @@ module Spreedly
       Transaction.new_from(xml_doc)
     end
 
-    def list_transactions(since_token = nil, payment_method_token = nil)
-      xml_doc = ssl_get(list_transactions_url(since_token, payment_method_token), headers)
+    def list_transactions(since_token = nil, payment_method_token = nil, options = {})
+      xml_doc = ssl_get(list_transactions_url(since_token, payment_method_token, options), headers)
       Transaction.new_list_from(xml_doc)
     end