+++ lib/spreedly/urls.rb
@@ -62,11 +62,20 @@ module Spreedly
       "#{base_url}/v1/gateways/#{gateway_token}/store.xml"
     end
 
-    def list_transactions_url(since_token, payment_method_token)
-      since_param = "?since_token=#{since_token}" if since_token
-      return "#{base_url}/v1/transactions.xml#{since_param}" unless payment_method_token
+    def list_transactions_url(since_token, payment_method_token, options = {})
+      options.each do |key, val|
+        options[key.to_sym] = val
+      end
 
-      "#{base_url}/v1/payment_methods/#{payment_method_token}/transactions.xml#{since_param}"
+      params = []
+      params << "since_token=#{since_token}" if since_token
+      params << "count=#{options[:count]}" if options[:count]
+      params << "order=#{options[:order]}" if options[:order]
+      params << "state=#{options[:state]}" if options[:state]
+      param_string = "?#{params.join('&')}" if params.any?
+      return "#{base_url}/v1/transactions.xml#{param_string}" unless payment_method_token
+
+      "#{base_url}/v1/payment_methods/#{payment_method_token}/transactions.xml#{param_string}"
     end
 
     def list_payment_methods_url(since_token)